Move api wsgi to api directory.
Move endpoints to __init__.py.
