+++ mash/services/api/wsgi.py
@@ -1,4 +1,4 @@
-from mash.services.api.endpoints import app as application
+from mash.services.api import app as application
 
 if __name__ == '__main__':
     application.run(port=5000)
